requirements.txt
asgiref==3.8.1
certifi==2025.1.31
channels==4.2.2
chardet==3.0.4
Django==5.2
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
Markdown==3.7
PyJWT==2.9.0
rfc3986==1.5.0
sniffio==1.3.1
sqlparse==0.5.3
tzdata==2025.2
